Restructure: Changed name to capital and moved text to own file
