+++ drunk-souls/src/rules/Rules.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import RulesText from './RulesText';
+
+class Rules extends React.Component {
+  render() {
+    document.title = "Drunk Souls - Rules";
+    return (
+      <RulesText/>
+    )
+  }
+}
+export default Rules
